requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain-community==0.3.29
langchain-experimental==0.3.4
langchain-neo4j==0.5.0
langchain-text-splitters==0.3.11
neo4j==5.28.2
python-dotenv==1.1.1
supabase==2.18.1
tiktoken==0.11.0
uvicorn==0.35.0
langchain-google-genai==2.1.10

# langchain
# langchain-google-genai
# langchain-community
# langchain-experimental
# langchain-neo4j
# neo4j
# python-dotenv
# tiktoken
# supabase
# fastapi
# uvicorn[standard]
# langchain-text-splitters